+++ src/main/java/com/spartafarmer/agri_commerce/AgriCommerceApplication.java
@@ -3,7 +3,6 @@ package com.spartafarmer.agri_commerce;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
 import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-import org.springframework.retry.annotation.EnableRetry;
 
 
 @EnableJpaAuditing
